HTTP/BooksAPI/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Marshmallow
marshmallow-sqlalchemy
psycopg2
python-dotenv
flask_migrate